[Tests] Split LinterTests into 3 files
